Many syntactic corrections (types mainly); patch from Frank Zago.

+++ branches/AMD64/ndiswrapper/driver/iw_ndis.c
@@ -873,8 +873,9 @@ static char *ndis_translate_scan(struct net_device *dev, char *event,
 				for (i = 0; i < ielen; i++)
 					p += sprintf(p, "%02x", iep[i]);
 				
-				DBGTRACE2("adding wpa_ie :%u",
-					  (u32)strlen(buf));
+				DBGTRACE2("adding wpa_ie :%lu",
+					  (unsigned long)strlen(buf));
+
 				memset(&iwe, 0, sizeof(iwe));
 				iwe.cmd = IWEVCUSTOM;
 				iwe.u.data.length = strlen(buf);
@@ -885,8 +886,8 @@ static char *ndis_translate_scan(struct net_device *dev, char *event,
 				for (i = 0; i < ielen; i++)
 					p += sprintf(p, "%02x", iep[i]);
 
-				DBGTRACE2("adding rsn_ie :%u\n",
-					  (u32)strlen(buf));
+				DBGTRACE2("adding rsn_ie :%lu\n",
+					  (unsigned long)strlen(buf));
 				memset(&iwe, 0, sizeof(iwe));
 				iwe.cmd = IWEVCUSTOM;
 				iwe.u.data.length = strlen(buf);
@@ -1287,18 +1288,16 @@ static int wpa_set_key(struct net_device *dev, struct iw_request_info *info,
 	}
 
 	if (wpa_key.key_len > sizeof(ndis_key.key)) {
-		DBGTRACE2("incorrect key length (%u)", (u32)wpa_key.key_len);
+		DBGTRACE2("incorrect key length (%lu)", (unsigned long)wpa_key.key_len);
 		TRACEEXIT2(return -1);
 	}
 	
 	if (wpa_key.seq_len > IW_ENCODING_TOKEN_MAX) {
-		DBGTRACE2("incorrect seq? length = (%u)",
-			  (u32)wpa_key.seq_len);
+		DBGTRACE2("incorrect seq? length = (%lu)", (unsigned long)wpa_key.seq_len);
 		TRACEEXIT2(return -1);
 	}
 
-	DBGTRACE2("setting key %d, %u", wpa_key.key_index,
-		  (u32)wpa_key.key_len);
+	DBGTRACE2("setting key %d, %lu", wpa_key.key_index, (unsigned long)wpa_key.key_len);
 	memset(&ndis_key, 0, sizeof(ndis_key));
 
 	ndis_key.struct_size = sizeof(ndis_key);
